More accurate rendering of {,Bad}String
Track which string delimiter was used when the string was parsed, and encode
non-printable and non-ascii characters in strings using hex escape codes.

The second change is necessary because we can't guarantee that the encoding
of the resulting file. Escaping code points is the most reasonable choice
here, given the constraints.